requirements.txt
blessed==1.22.0
blinker==1.9.0
myst-parser==4.0.1
feedgenerator==2.2.1
inquirer==3.4.1
Jinja2==3.1.6
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.3
minchin.pelican.plugins.nojekyll==1.2.0
mdurl==0.1.2
pelican==4.11.0
pelican-liquid-tags==1.0.4
pelican-sitemap==1.2.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-editor==1.0.4
python-slugify==8.0.4
pytz==2025.2
readchar==4.2.1
rich==14.2.0
six==1.17.0
smartypants==2.0.2
text-unidecode==1.3
tzdata==2025.2
typogrify==2.1.0
Unidecode==1.4.0
wcwidth==0.2.14
pyyaml
duty
pelican-related-posts
pelican-neighbors
pelican-seo
livereload
